Queue skill button requests so they appear in the UI
When a user clicks a skill button (e.g. 'Create Pull Request', 'Merge Changes')
while the agent is busy processing a previous request, the new request was silently
rejected by sendRequest because hasPendingRequest was true and no queue option was set.
The request was never added to the model, so nothing appeared in the chat UI.

Fix: pass queue: ChatRequestQueueKind.Queued in both code paths:
- chatSessions.contribution.ts (OpenChatSessionAction, used by Copilot extension buttons)
- agentHostSkillButtons.ts (VS Code native skill button actions)

Also extracts a shared resolvePromptSlashCommand.ts helper used by both paths
to resolve slash commands to prompt file variable entries for context attachment.

(Written by Copilot)
