+++ src/vs/workbench/contrib/chat/browser/chatSessions/chatSessions.contribution.ts
@@ -35,9 +35,9 @@ import { ChatSessionOptionsMap, ChatSessionStatus, IChatNewSessionRequest, IChat
 import { ChatAgentLocation, ChatModeKind } from '../../common/constants.js';
 import { CHAT_CATEGORY } from '../actions/chatActions.js';
 import { IChatEditorOptions } from '../widgetHosts/editor/chatEditor.js';
-import { IChatService, ResponseModelState } from '../../common/chatService/chatService.js';
+import { ChatRequestQueueKind, IChatService, ResponseModelState } from '../../common/chatService/chatService.js';
 import { autorun, observableFromEvent } from '../../../../../base/common/observable.js';
-import { IChatRequestVariableEntry, PromptFileVariableKind, toPromptFileVariableEntry } from '../../common/attachments/chatVariableEntries.js';
+import { IChatRequestVariableEntry } from '../../common/attachments/chatVariableEntries.js';
 import { IViewsService } from '../../../../services/views/common/viewsService.js';
 import { ChatViewId } from '../chat.js';
 import { ChatViewPane } from '../widgetHosts/viewPane/chatViewPane.js';
@@ -48,11 +48,10 @@ import { isUntitledChatSession, LocalChatSessionUri } from '../../common/model/c
 import { assertNever } from '../../../../../base/common/assert.js';
 import { ICommandService } from '../../../../../platform/commands/common/commands.js';
 import { Target } from '../../common/promptSyntax/promptTypes.js';
-import { slashReg } from '../../common/requestParser/chatRequestParser.js';
-import { OffsetRange } from '../../../../../editor/common/core/ranges/offsetRange.js';
 import { ILanguageModelToolsService } from '../../common/tools/languageModelToolsService.js';
 import { IChatModel } from '../../common/model/chatModel.js';
 import { ICustomizationHarnessService } from '../../common/customizationHarnessService.js';
+import { resolvePromptSlashCommandToVariableEntry } from '../../common/promptSyntax/resolvePromptSlashCommand.js';
 
 const extensionPoint = ExtensionsRegistry.registerExtensionPoint<IChatSessionsExtensionPoint[]>({
 	extensionPoint: 'chatSessions',
@@ -567,12 +566,12 @@ export class ChatSessionsService extends Disposable implements IChatSessionsServ
 						const resource = URI.revive(chatOptions.resource);
 						const ref = await chatService.acquireOrLoadSession(resource, ChatAgentLocation.Chat, CancellationToken.None, 'ChatSessionsContribution#sendPrompt');
 						try {
-							const promptFile = await resolvePromptSlashCommand(chatOptions.prompt, contribution.type, customizationHarnessService, toolsService);
+							const promptFile = await resolvePromptSlashCommandToVariableEntry(chatOptions.prompt, contribution.type, customizationHarnessService, toolsService, CancellationToken.None);
 							if (promptFile) {
 								attachedContext = [promptFile, ...(attachedContext ?? [])];
 							}
 
-							const result = await chatService.sendRequest(resource, chatOptions.prompt, { agentIdSilent: type, attachedContext });
+							const result = await chatService.sendRequest(resource, chatOptions.prompt, { agentIdSilent: type, attachedContext, queue: ChatRequestQueueKind.Queued });
 							if (result.kind === 'queued') {
 								await result.deferred;
 							} else if (result.kind === 'sent') {
@@ -1383,7 +1382,7 @@ async function openChatSession(accessor: ServicesAccessor, openOptions: NewChatS
 			}
 
 			let attachedContext = chatSendOptions.attachedContext;
-			const promptFile = await resolvePromptSlashCommand(chatSendOptions.prompt, openOptions.type, customizationHarnessService, toolsService);
+			const promptFile = await resolvePromptSlashCommandToVariableEntry(chatSendOptions.prompt, openOptions.type, customizationHarnessService, toolsService, CancellationToken.None);
 			if (promptFile) {
 				attachedContext = [promptFile, ...(attachedContext ?? [])];
 			}
@@ -1414,26 +1413,6 @@ function normalizeSessionOptions(options: ReadonlyChatSessionOptionsMap | Readon
 	}
 	// Plain object fallback (e.g. from JSON deserialization)
 	return ChatSessionOptionsMap.fromRecord(options as unknown as Record<string, string | IChatSessionProviderOptionItem>);
-}
-
-/**
- * Returns the variable entry for a slash command if the prompt starts with a slash command that can be resolved to a prompt file, otherwise returns undefined.
- */
-async function resolvePromptSlashCommand(prompt: string, sessionType: string, customizationHarnessService: ICustomizationHarnessService, toolsService: ILanguageModelToolsService): Promise<IChatRequestVariableEntry | undefined> {
-	const slashMatch = prompt.match(slashReg);
-	// starts with a slash command, add the corresponding prompt file to the context if it exists
-	if (slashMatch) {
-		// need to resolve the slash command to get the prompt file
-		const slashCommand = await customizationHarnessService.resolvePromptSlashCommand(slashMatch[1], sessionType, CancellationToken.None);
-		if (slashCommand) {
-			const parseResult = slashCommand.parsedPromptFile;
-			// add the prompt file to the context
-			const refs = parseResult.body?.variableReferences.map(({ name, offset, fullLength }) => ({ name, range: new OffsetRange(offset, offset + fullLength) })) ?? [];
-			const toolReferences = toolsService.toToolReferences(refs);
-			return toPromptFileVariableEntry(parseResult.uri, PromptFileVariableKind.PromptFile, undefined, true, toolReferences);
-		}
-	}
-	return undefined;
 }
 
 export function getResourceForNewChatSession(options: NewChatSessionOpenOptions): URI {
